externalize resolver
