--- packages/svelte-playwright-coverage/src/spc/analyze/resolve/fetchSource.ts
@@ -1,11 +1,0 @@
-import { readFile } from 'node:fs/promises'
-
-export const fetchSource = async (url: string) => {
-  try {
-    const source = await readFile(new URL(url), 'utf-8')
-
-    return source
-  } catch {
-    return null
-  }
-}
