--- packages/svelte-playwright-coverage/src/spc/analyze/resolve/index.ts
@@ -1,6 +1,0 @@
-import { resolvePlaywright } from './resolvePlaywright.js'
-import { resolveVite } from './resolveVite.js'
-
-export const resolve = async () => {
-  await Promise.all([resolvePlaywright(), resolveVite()])
-}
