--- packages/svelte-playwright-coverage/src/spc/analyze/resolve/pickSourceMappingURL.ts
@@ -1,2 +1,0 @@
-export const pickSourceMappingURL = (source: string | undefined | null) =>
-  source?.match(/\/\/# sourceMappingURL=(.*)/)?.[1]
