--- packages/svelte-playwright-coverage/src/types/AwaitedReturn.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type AwaitedReturn<T extends (...args: any[]) => Promise<any>> = Awaited<
-  ReturnType<T>
->
