--- packages/svelte-playwright-coverage/src/types/PlaywrightV8RawCoverage.ts
@@ -1,6 +1,0 @@
-import { Page } from '@playwright/test'
-import { AwaitedReturn } from './AwaitedReturn.js'
-
-export type PlaywrightV8RawCoverage = AwaitedReturn<
-  Page['coverage']['stopJSCoverage']
->
